% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm.R
\name{dmvnorm}
\alias{dmvnorm}
\title{The multivariate normal density function.}
\usage{
dmvnorm(x, mu, sigma)
}
\arguments{
\item{x}{A p-dimension vector.}

\item{mu}{A p-dimensional mean vector.}

\item{sigma}{A p \times p variance-covariance matrix.}
}
\value{
A scalar density at \code{x} for the distribution defined by \code{mu} and \code{sigma}.
}
\description{
The multivariate normal density function.
}
