% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Generate random vectors from a multivariate normal distribution.}
\usage{
rmvnorm(n, mu, sigma)
}
\arguments{
\item{n}{Number of random vectors to return.}

\item{mu}{The mean vector.}

\item{sigma}{The variance-covariance matrix.}
}
\value{
A matrix of \code{n} rows, each drawn from the MVN distribution with parameters \code{mu} and \code{sigma}.
}
\description{
Generate random vectors from a multivariate normal distribution.
}
