% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{subsets}
\alias{subsets}
\title{List all subsets of a set.}
\usage{
subsets(x, minimum = 1, maximum = length(x))
}
\arguments{
\item{x}{A set-like object (coercable to a vector or list).}

\item{minimum}{The smallest subset to list.}

\item{maximum}{The largest subset to list.}
}
\value{
A list of subsets, each a set.
}
\description{
List all subsets of a set.
}
