% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_output.R
\name{teXtab}
\alias{teXtab}
\title{Generate a LaTeX tabular enironment from a rectangular R object.}
\usage{
teXtab(A, latex_env = "bmatrix", .digits = NULL, .margin = 1)
}
\arguments{
\item{A}{A matrix, dataframe or similar.}

\item{latex_env}{The type of LaTeX tabular environment to generate.}

\item{.digits}{Optional digits to round table elements.}

\item{.margin}{The dimension to iterate over - will be output as rows.}
}
\description{
Generate a LaTeX tabular enironment from a rectangular R object.
}
